consolidate all time series extension code
