+++ snuba/query/timeseries_extension.py
@@ -1,19 +1,59 @@
 from datetime import datetime, timedelta
 from typing import Any, Mapping, Tuple
 
-from snuba import state
-from snuba.util import parse_datetime
+from snuba import environment, state
 from snuba.query.conditions import (
-    binary_condition,
     BooleanFunctions,
     ConditionFunctions,
+    binary_condition,
 )
 from snuba.query.expressions import Column, Literal
 from snuba.query.extensions import QueryExtension
 from snuba.query.logical import Query
 from snuba.query.processors import ExtensionData, ExtensionQueryProcessor
 from snuba.request.request_settings import RequestSettings
-from snuba.schemas import get_time_series_extension_properties
+from snuba.util import parse_datetime
+from snuba.utils.metrics.backends.wrapper import MetricsWrapper
+from snuba.utils.metrics.decorators import track_calls
+
+
+timeseries_metrics = MetricsWrapper(environment.metrics, "extensions.timeseries")
+
+
+def get_time_series_extension_properties(
+    default_granularity: int, default_window: timedelta
+):
+    return {
+        "type": "object",
+        "properties": {
+            "from_date": {
+                "type": "string",
+                "format": "date-time",
+                "default": track_calls(
+                    timeseries_metrics,
+                    "from_date.default",
+                    lambda: (
+                        datetime.utcnow().replace(microsecond=0) - default_window
+                    ).isoformat(),
+                ),
+            },
+            "to_date": {
+                "type": "string",
+                "format": "date-time",
+                "default": track_calls(
+                    timeseries_metrics,
+                    "to_date.default",
+                    lambda: datetime.utcnow().replace(microsecond=0).isoformat(),
+                ),
+            },
+            "granularity": {
+                "type": "number",
+                "default": default_granularity,
+                "minimum": 1,
+            },
+        },
+        "additionalProperties": False,
+    }
 
 
 class TimeSeriesExtensionProcessor(ExtensionQueryProcessor):
